% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multicross.R
\name{matchpar}
\alias{matchpar}
\title{This function takes an scRNAseq counts matrix as an input (cells in rows x genes in cols) and outputs a matrix of cells x 3 covariates (number of genes detected, sequencing depth and mitochondrial gene expression). This covariate matrix can then be used to match cells and perform stratified permutation.}
\usage{
matchpar(datamat)
}
\arguments{
\item{datamat}{is cells x genes matrix from an scRNAseq experiment. Could be non-UMI or UMI counts.}
}
\value{
A list of scRNAseq matrices subsetted by KEGG Pathways, so list[[n]] corresponds to the n^th pathway.
}
\description{
This function takes an scRNAseq counts matrix as an input (cells in rows x genes in cols) and outputs a matrix of cells x 3 covariates (number of genes detected, sequencing depth and mitochondrial gene expression). This covariate matrix can then be used to match cells and perform stratified permutation.
}
