% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multicross.R
\name{mhcccreate}
\alias{mhcccreate}
\title{Creates the null covariance matrix for Mmcm, corresponding to the scenario when all K distributions are the same}
\usage{
mhcccreate(nvec)
}
\arguments{
\item{nvec}{is a vector containing the sizes of the K different classes}
}
\value{
The inputs for the Multisample Mahalanobis Crossmatch Test
}
\description{
Creates the null covariance matrix for Mmcm, corresponding to the scenario when all K distributions are the same
}
