% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multicross.R
\name{mhccexecutelong}
\alias{mhccexecutelong}
\title{Calculates the pairwise crosscounts for the K classes being examined}
\usage{
mhccexecutelong(nvec, apmat)
}
\arguments{
\item{nvec}{is a vector containing the sizes of the K different classes}

\item{apmat}{is the data matrix containing pooled data from each of the K classess, on which optimal non-bipartite matching is performed.}
}
\value{
The inputs for the Multisample Mahalanobis Crossmatch Tests
}
\description{
Calculates the pairwise crosscounts for the K classes being examined
}
