% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multicross.R
\name{multigene}
\alias{multigene}
\title{Given two input matrices with the same number of observations but differrent number of variables, this function returns the largest canonical correlation between variables of matrix 1 (X) and those of matrix 2 (Y).}
\usage{
multigene(X, Y)
}
\arguments{
\item{X}{is a data matrix with observations as rows and features/genes as columns.}

\item{Y}{is a data matrix with observations as rows (same observations as X) and a different set of features/genes as columns.}
}
\value{
The largest canonical correlation between X and Y as described in <https://ncss-wpengine.netdna-ssl.com/wp-content/themes/ncss/pdf/Procedures/NCSS/Canonical_Correlation.pdf>.
}
\description{
Given two input matrices with the same number of observations but differrent number of variables, this function returns the largest canonical correlation between variables of matrix 1 (X) and those of matrix 2 (Y).
}
